Add product feature implelemt

+++ DevSlkill.Inventory/DevSkill.Inventory.Web/Areas/Admin/Models/UpdateProductModel.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string ProductName { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public int Price { get; set; }
-        public int Ratings { get; set; }
+        public double Price { get; set; }
+        public double Ratings { get; set; }
     }
 }
